src/auth/requirements.txt
pylint
jedi
pyjwt
flask
flask_mysqldb
wheel 
setuptools
